+++ packages/ui/src/routes/workspaceRoutes.tsx
@@ -8,6 +8,7 @@ import { WorkforceRoute } from "@ui/routes/WorkforceRoute";
 import { ZoneDetailRoute } from "@ui/routes/ZoneDetailRoute";
 import { StrainsRoute } from "@ui/routes/StrainsRoute";
 import { StructuresRoute } from "@ui/routes/StructuresRoute";
+import { StructureRoute } from "@ui/routes/StructureRoute";
 
 export const workspaceRoutes = createRoutesFromElements(
   <Route
@@ -28,6 +29,7 @@ export const workspaceRoutes = createRoutesFromElements(
     <Route index element={<Navigate to={workspaceTopLevelRoutes.company.path} replace />} />
     <Route path={workspaceTopLevelRoutes.company.path} element={<DashboardRoute />} />
     <Route path={workspaceTopLevelRoutes.structures.path} element={<StructuresRoute />} />
+    <Route path="structures/:structureId" element={<StructureRoute />} />
     <Route path="structures/:structureId/zones/:zoneId" element={<ZoneDetailRoute />} />
     <Route path={workspaceTopLevelRoutes.hr.path} element={<WorkforceRoute />} />
     <Route path={workspaceTopLevelRoutes.strains.path} element={<StrainsRoute />} />